Reorganize utils module
